+++ examples/007-clean_architecture_dane_mackier_app/lib/ui/pages/post_page/like_button.dart
@@ -12,7 +12,7 @@ class LikeButton extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     //NOTE1: get reactiveModel of PostsService
-    final postsServiceRM = Injector.getAsReactive<PostsService>();
+    final postsServiceRM = ReactiveModel<PostsService>();
 
     return Row(
       children: <Widget>[